// src/shared/ui/components/utility/ListMenu/interfaces/IMenuListStyleProps.ts
/**
 * MenuListStyleProps Interface
 * 
 * Defines style properties that can be applied to the ListMenu component
 * for customization and theming.
 * 
 * @interface IMenuListStyleProps
 */

/**
 * Style properties interface for ListMenu component
 */
export interface IMenuListStyleProps {
  /** CSS position property for menu */
  position?: string;
  
  /** Width of the menu */
  width?: string;
  
  /** Font size for menu items */
  fontSize?: string;
  
  /** Font weight for menu items */
  fontWeight?: string;
  
  /** Border radius for menu */
  radius?: string;
  
  /** Size of menu icon */
  iconSize?: string;
  
  /** Padding within the menu */
  padding?: string;
  
  /** CSS display property for menu */
  display?: string;
}